backend: (llvm) move convert_ctypes to xdsl.jit.llvm
Sasha's suggestion: hang JIT infrastructure off a top-level xdsl.jit
package so future targets (WASM, in-xDSL JIT) don't have to live under
xdsl.backend. The LLVM-specific layer becomes xdsl.jit.llvm.
